My last commit
This is my last commit (unless huuuuuuuuuge bugs appear in the next 10
minutes) on the adsabs repo as ADS employee.
I don't exclude to contribute in the future, but  I'll not be an
employee any more.

This project has probably been the best work I've done so far and I'm
proud of it.

Jay, Roman, Alex, Rahul, Edwin: take good care of it!